Refactor class structure
