+++ src/com/lucasallegri/datdec/compile/Compile.java
@@ -1,4 +1,4 @@
-package com.lucasallegri.decompile;
+package com.lucasallegri.datdec.compile;
 
 import java.io.File;
 import java.io.FileInputStream;
@@ -6,16 +6,16 @@ import java.io.FileOutputStream;
 import java.io.IOException;
 import java.util.List;
 
-import com.lucasallegri.gui.DatdecContext;
-import com.lucasallegri.gui.DatdecGUI;
-import com.lucasallegri.ooo.OOOExporter;
-import com.lucasallegri.util.ConfigChoice;
-import com.lucasallegri.util.FileUtil;
-import com.lucasallegri.util.PathUtil;
+import com.lucasallegri.datdec.gui.DatdecContext;
+import com.lucasallegri.datdec.gui.DatdecGUI;
+import com.lucasallegri.datdec.ooo.OOOImporter;
+import com.lucasallegri.datdec.util.ConfigChoice;
+import com.lucasallegri.datdec.util.FileUtil;
+import com.lucasallegri.datdec.util.PathUtil;
 
-public class Decompile {
+public class Compile {
 	
-	public static void decompile() {
+	public static void compile() {
 		
 		File source = new File(PathUtil.getPathToConfig(DatdecContext.selectedConfig));
 		FileInputStream in = null;
@@ -24,25 +24,26 @@ public class Decompile {
 		try {
 			
 			in = new FileInputStream(source);
-			String dest = path.replaceFirst("\\.dat$", ".xml");
+			String dest = path.replaceFirst("\\.xml$", ".dat");
+			if(DatdecContext.doBackups) FileUtil.backupFile(dest);
 			FileOutputStream out = new FileOutputStream(dest);
-			OOOExporter.export(in, out);
-			DatdecGUI.pushState(DatdecContext.selectedConfig + " was decompiled.");
+			OOOImporter._import(in, out);
+			source.delete();
+			DatdecGUI.pushState(DatdecContext.selectedConfig + " was compiled.");
 			ConfigChoice.populate();
 			
 		} catch (IOException e) {
 			e.printStackTrace();
 		}
-		
 	}
 	
-	public static void decompileAll() {
+	public static void compileAll() {
 		
 		List<String> _list = FileUtil.fileNamesInDirectory("rsrc/config/");
 		
 		for(int i = 0; i < _list.size(); i++) {
 			
-			if(_list.get(i).endsWith(".xml")) continue;
+			if(_list.get(i).endsWith(".dat")) continue;
 			
 			File source = new File(PathUtil.getPathToConfig(_list.get(i)));
 			FileInputStream in = null;
@@ -51,16 +52,18 @@ public class Decompile {
 			try {
 				
 				in = new FileInputStream(source);
-				String dest = path.replaceFirst("\\.dat$", ".xml");
+				String dest = path.replaceFirst("\\.xml$", ".dat");
+				if(DatdecContext.doBackups) FileUtil.backupFile(dest);
 				FileOutputStream out = new FileOutputStream(dest);
-				OOOExporter.export(in, out);
+				OOOImporter._import(in, out);
+				source.delete();
 				
 			} catch (IOException e) {
 				e.printStackTrace();
 			}
 		}
 		
-		DatdecGUI.pushState(_list.size() + " configs were decompiled.");
+		DatdecGUI.pushState(_list.size() + " configs were compiled.");
 		ConfigChoice.populate();
 		
 	}
