+++ src/com/lucasallegri/datdec/decompile/Decompile.java
@@ -1,4 +1,4 @@
-package com.lucasallegri.compile;
+package com.lucasallegri.datdec.decompile;
 
 import java.io.File;
 import java.io.FileInputStream;
@@ -6,16 +6,16 @@ import java.io.FileOutputStream;
 import java.io.IOException;
 import java.util.List;
 
-import com.lucasallegri.gui.DatdecContext;
-import com.lucasallegri.gui.DatdecGUI;
-import com.lucasallegri.ooo.OOOImporter;
-import com.lucasallegri.util.ConfigChoice;
-import com.lucasallegri.util.FileUtil;
-import com.lucasallegri.util.PathUtil;
+import com.lucasallegri.datdec.gui.DatdecContext;
+import com.lucasallegri.datdec.gui.DatdecGUI;
+import com.lucasallegri.datdec.ooo.OOOExporter;
+import com.lucasallegri.datdec.util.ConfigChoice;
+import com.lucasallegri.datdec.util.FileUtil;
+import com.lucasallegri.datdec.util.PathUtil;
 
-public class Compile {
+public class Decompile {
 	
-	public static void compile() {
+	public static void decompile() {
 		
 		File source = new File(PathUtil.getPathToConfig(DatdecContext.selectedConfig));
 		FileInputStream in = null;
@@ -24,26 +24,25 @@ public class Compile {
 		try {
 			
 			in = new FileInputStream(source);
-			String dest = path.replaceFirst("\\.xml$", ".dat");
-			if(DatdecContext.doBackups) FileUtil.backupFile(dest);
+			String dest = path.replaceFirst("\\.dat$", ".xml");
 			FileOutputStream out = new FileOutputStream(dest);
-			OOOImporter._import(in, out);
-			source.delete();
-			DatdecGUI.pushState(DatdecContext.selectedConfig + " was compiled.");
+			OOOExporter.export(in, out);
+			DatdecGUI.pushState(DatdecContext.selectedConfig + " was decompiled.");
 			ConfigChoice.populate();
 			
 		} catch (IOException e) {
 			e.printStackTrace();
 		}
+		
 	}
 	
-	public static void compileAll() {
+	public static void decompileAll() {
 		
 		List<String> _list = FileUtil.fileNamesInDirectory("rsrc/config/");
 		
 		for(int i = 0; i < _list.size(); i++) {
 			
-			if(_list.get(i).endsWith(".dat")) continue;
+			if(_list.get(i).endsWith(".xml")) continue;
 			
 			File source = new File(PathUtil.getPathToConfig(_list.get(i)));
 			FileInputStream in = null;
@@ -52,18 +51,16 @@ public class Compile {
 			try {
 				
 				in = new FileInputStream(source);
-				String dest = path.replaceFirst("\\.xml$", ".dat");
-				if(DatdecContext.doBackups) FileUtil.backupFile(dest);
+				String dest = path.replaceFirst("\\.dat$", ".xml");
 				FileOutputStream out = new FileOutputStream(dest);
-				OOOImporter._import(in, out);
-				source.delete();
+				OOOExporter.export(in, out);
 				
 			} catch (IOException e) {
 				e.printStackTrace();
 			}
 		}
 		
-		DatdecGUI.pushState(_list.size() + " configs were compiled.");
+		DatdecGUI.pushState(_list.size() + " configs were decompiled.");
 		ConfigChoice.populate();
 		
 	}
